requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
numpy==1.24.4
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
pillow==10.2.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
